Move http helpers into httputil.go
